+++ pretraining/train.py
@@ -20,7 +20,7 @@ def loss_fn(model_state, model_graphdef, x, pad=False): # [B, T]
     loss_mask = data.pad_mask(x) if pad else jnp.ones(x.shape, dtype=bool)
     loss_mask = loss_mask.at[:, -1].set(False)
     logits = model(x) # [B, T, V]
-    losses = optax.softmax_cross_entropy_with_integer_labels(logits, y) # [B, T]
+    losses = optax.softmax_cross_entropy_with_integer_labels(logits.astype(jnp.float32), y) # [B, T]
     return (losses * loss_mask).sum() / loss_mask.sum()
